seed_script新規追加

+++ backend/app/models/pulse_survey.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Column, Integer, ForeignKey, Text, DateTime, Date, UniqueConstraint, CheckConstraint
+from sqlalchemy import Column, Integer, ForeignKey, Text, DateTime, Date, UniqueConstraint, CheckConstraint, Float
 from sqlalchemy.orm import relationship
 from datetime import datetime
 
@@ -12,7 +12,7 @@ class PulseSurvey(Base):
 
     user_id = Column(Integer, ForeignKey("users.id"), nullable=False)
 
-    score = Column(Integer, CheckConstraint('score >= 1 AND score <= 5'), nullable=False)
+    score = Column(Float, CheckConstraint('score >= 1 AND score <= 5'), nullable=False)
     
     memo = Column(Text, nullable=True) # 空白も許容
